Encapsulated rendering logic into the renderer namespace
