+++ src/renderer.hpp
@@ -4,15 +4,18 @@
 
 #ifndef RENDERER_HPP
 #define RENDERER_HPP
+#include "bgfx/bgfx.h"
 
-namespace Abyss
+/// \brief The renderer singleton.
+/// \details This singleton is used to handle BGFX rendering, and the rendering loop.
+/// All Windowing (GLFW) details don't belong here.
+namespace Abyss::renderer
 {
+    int init(bgfx::Init init);
 
-class renderer
-{
+    void render(int width, int height);
 
-};
-
+    void reset();
 } // Abyss
 
 #endif //RENDERER_HPP
